Refactor: MainPage now in app router tree
Now we can use routing in main page.

still have some issues.

1) while on rank matching, even we route.push to profile, the matching
   router's matching status didn't change. seems like some kind of
   caching is ongoing. need to deal with it.

2) when do conditional render on match router, it warns about state
   change. need to check it.
