+++ front/src/app/main_frame/page.tsx
@@ -2,15 +2,13 @@
 
 import { useState, useEffect } from 'react';
 
-// import Main from './containers';
-
-import Mainbox from '@/components/mainbox/mainbox';
+import Matching from './match/page';
 
 export default function MainFrame () {
 
   return (
-    // <Mainbox mod={clicked} search={id} setProfile={setProfile}/>
-      <Mainbox />
+
+      <Matching />
 
   );
 }
