+++ front/src/components/profile/myprofile.module.css
@@ -1,0 +1,7 @@
+.myprofileimg
+{
+  position: absolute;
+  width: 100%;
+  height: 29%;
+  max-height: 30%;
+}
